+++ perSystem/packages.nix
@@ -2,16 +2,15 @@
   perSystem = { hsPkgs, ... }:
     let
       tda = hsPkgs.hermod-tracing-api;
-      td  = hsPkgs.trace-dispatcher;
+      td  = hsPkgs.hermod-tracing-core;
       hrf = hsPkgs.hermod-recon-framework;
       htr = hsPkgs.hermod-trace-resources;
     in
     {
       packages.hermod-tracing-api             = tda.components.library;
 
-
-      packages.trace-dispatcher              = td.components.library;
-      checks.trace-dispatcher-test           = td.components.tests.trace-dispatcher-test;
+      packages.hermod-tracing-core           = td.components.library;
+      checks.hermod-tracing-core-test        = td.components.tests.hermod-tracing-core-test;
 
       packages.hermod-recon                  = hrf.components.exes.hermod-recon;
       packages.hermod-recon-grep             = hrf.components.exes.hermod-recon-grep;